requirements.txt
python-swiftclient
python-keystoneclient
bottle
pycrypto
